Create contract_setup module
